+++ kotlin/lib/src/main/kotlin/models/MessagePrecheckOut.kt
@@ -1,0 +1,10 @@
+// This file is @generated
+package com.svix.kotlin.models
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class MessagePrecheckOut(
+    /** Whether there are any active endpoint that would get sent such a message. */
+    val active: Boolean
+)